Minor style updates
Color changes to bring it closer to [maas.io](http://maas.io), which was the first similar site of ours I found.  I'm not sure if there are specific guidelines for these styles.

I still think the Request a Review call-to-action is a bit subtle being up there in the header, but it's at least a bit better this way.