Handle charms with terms.
In order to download charms with terms, we need to be auth'd to the
charmstore, and we need to agree to the charm terms before
downloading.

A charmstore user/pass will be passed in via charm config. We then
use pexpect to `charm login` dynamically as needed. Any
other method would be problematic since the charmstore macaroons
and sso token will eventually expire, making it necessary to login
again.

Also added a catch-all 500 view just for the sake of aesthetics.