Surface 'charm pull' errors.
Even when the charmstore api responds successfully to a revision url, there can be
subsequent problems retrieving the source for that revision using 'charm pull'.
Some of these can (and will) be handled gracefully by the app - for example, agreeing
to terms. But some require user intervention - for example, when the user has
failed to 'charm release-term' all the terms required.

This change surfaces such problems (specifically, any problem that occurs during a
'charm pull' or 'juju agree' that can not be automatically handled) to the user by
showing an error message on the review page (instead of the unhelpful 'Oops' message),
along with instructions for resolving the problem when possible.

Fixes #69.